the new shared edit form. a dynamic form that allows for extensions without overwriting this view.
By adding a 'group' to the model column options fields are made in to sets , with the value of group being the set name.

The set name is then used as the tab name in the front end and a url&underscored version of the name is then used to check for partials. 

As the view stack applies to partials you can overwrite a group partial as you want to within you application.

It should also allow for a method to handle join interfaces as well.