new mapping functions to fix url when updating content.
the url is linked to the content item its made on and is inherited by any drafts.

when reverting to a previous version new urls/joins in general are not copied over to be consistent.

might need to add a layering on this so if reverting to a previous version that has missing urls those are redirected rather than ignored. for now i'll add a warning message.